2.25 #2
timearray 분리
